server/fastapi/app/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
supabase==1.0.3
python-dotenv==1.0.0

